requirements.txt
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
transformers==4.36.0
torch==2.1.0
openai-whisper==20250625
ffmpeg-python==0.2.0
streamlit==1.29.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
plotly==5.17.0
